rework site header
reworks site header to reactively update based on the current user's role as defined on login using UserContext.js. The site header will display the same four buttons if the user is not an admin, but in the case that they are an admin, will replace the registration button with 'ADMIN DASH' linking to the admin dashboard. It will also re-link the scheduling button to '/adminscheduling'.